requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
pyyaml
jupyterlab>=3.5.0
ipywidgets>=7.7.0
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0

